Extract generic view TitleAndSubtitleRow from ItemToFulfillRow
