+++ WooCommerce/Classes/ViewRelated/ReusableViews/SwiftUI Components/ItemToFulfillRow.swift
@@ -8,26 +8,7 @@ struct ItemToFulfillRow: View, Identifiable {
     let subtitle: String
 
     var body: some View {
-        HStack {
-            VStack(alignment: .leading,
-                   spacing: 8) {
-                Text(title)
-                    .bodyStyle()
-                Text(subtitle)
-                    .footnoteStyle()
-            }.padding([.leading, .trailing], Constants.vStackPadding)
-            Spacer()
-        }
-        .padding([.top, .bottom], Constants.hStackPadding)
-        .frame(minHeight: Constants.height)
-    }
-}
-
-private extension ItemToFulfillRow {
-    enum Constants {
-        static let vStackPadding: CGFloat = 16
-        static let hStackPadding: CGFloat = 10
-        static let height: CGFloat = 64
+        TitleAndSubtitleRow(title: title, subtitle: subtitle)
     }
 }
